Changed strutcture of project
