--- index.html
@@ -1,27 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Fit File Analysis Tool</title>
-</head>
-<body>
-    <form id="upload-form" method="post" action="main.py/upload_file/" enctype="multipart/form-data">
-        <label for="fit-file" class="upload">Upload .fit file here</label>
-        <input type="file" id="fit-file" name="fit-file" accept=".fit">
-        <input type="submit" value="Submit">
-    </form>
-
-    <script>
-        async function display() {
-            const response = await fetch ('http://127.0.0.1:8000/decode_file/')
-            if (!response.ok) {
-                throw new Error("Network response was not ok")
-            }
-            const data = await repsonse.json()
-            console.log(data);
-        }
-        display()
-    </script>
-</body>
-</html>
