+++ utils/fit_csv_util.py
@@ -1,0 +1,14 @@
+csv = []
+
+def csv_head(record_fields):
+    header = []
+    header.append(record_fields)
+    csv.append(header)
+    return csv
+
+
+def csv_body(record_data):
+    body = []
+    body.append(record_data)
+    csv.append(body)
+    return csv
